15 min session Authorized user access hidden folders Actual ip being printed instead of nginx ip(localhost) Firebase refactoring
