+++ server/nginx-local.conf
@@ -1,6 +1,6 @@
 server {
 	listen 443 ssl;
-	
+
 	ssl_certificate /home/pi/cert/bundle_cert.crt;
 	ssl_certificate_key /home/pi/cert/bundle_cert.key;
 
@@ -12,6 +12,7 @@ server {
 
 	location / {
 		proxy_set_header X-Forwarded-Proto $scheme;
+		proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
 		proxy_pass http://localhost:8000;
 	}
 }
