+++ server/routes/files.js
@@ -103,6 +103,12 @@ router.get('/get/*', function(req, res) {
         var obj = [];
 
         for (var file of files) {
+
+          if (file.substring(0,1) === '.') {
+            if (!req.user) {
+              continue;
+            }
+          }
           const stat = fs.statSync(filePath+'/'+file);
           obj.push({
             'name': file,
